feat(Homepage,pages): inital draft for project-level report page
splitting server into homepage and project report page along with setting up email alerts charts
