+++ dashboard/__init__.py
@@ -61,9 +61,8 @@ if PROCESSOR_DB_URI is None:
     logger.warning("  ❌ ️No PROCESSOR_DB_URI is specified")
     sys.exit(1)
 
-"""
+
 ALERTS_DB_URI = os.environ.get('ALERTS_DB_URI', None)
 if ALERTS_DB_URI is None:
     logger.warning("  ❌ ️No ALERTS_DB_URI is specified")
     sys.exit(1)
-"""
